etcdserver/api/v2discovery: move internal "discovery" package
