+++ etcdserver/api/v2discovery/discovery_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package discovery
+package v2discovery
 
 import (
 	"context"
